requirements.txt
aioredis==1.3.1
amqp==2.6.1
appdirs==1.4.3
asgiref==3.2.10
asn1crypto==0.24.0
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
autopep8==1.5.3
Babel==2.8.0
beautifulsoup4==4.9.0
billiard==3.6.3.0
black==19.10b0
bleach==3.1.5
bump2version==1.0.0
bumpversion==0.6.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.0
chardet==3.0.4
check-manifest==0.42
click==7.1.1
colorama==0.4.3
constantly==15.1.0
coverage==5.1
cryptography==2.9.2
daphne==2.5.0
distlib==0.3.1
Django==3.0.5
django-bootstrap-modal-forms==1.5.0
django-bootstrap4==1.1.1
django-celery-beat==2.0.0
django-channels==0.7.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-rest-framework==0.1.0
django-timezone-field==4.0
django-user-agents==0.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docker==4.2.2
docutils==0.16
entrypoints==0.3
factory-boy==2.11.1
Faker==4.1.1
filelock==3.0.12
flake8==3.7.9
gunicorn==20.0.4
hiredis==1.0.1
httpie==2.2.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
jsonfield==3.1.0
keyring==21.2.1
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
model-mommy==2.0.0
more-itertools==8.4.0
msgpack==1.0.0
mysqlclient
nodejs==0.1.1
oauthlib==3.1.0
optional-django==0.1.0
packaging==20.4
pathspec==0.8.0
pep517==0.8.2
phonenumbers==8.12.2
Pillow==7.2.0
pkginfo==1.5.0.1
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypiwin32==223
PySocks==1.7.1
pytest==4.5.0
pytest-asyncio==0.14.0
pytest-cov==2.7.1
pytest-django==3.4.8
pytest-pythonpath==0.7.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-memcached==1.59
pytz==2019.3
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.3.1
readme-renderer==26.0
redis==3.5.3
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sockjs-tornado==1.0.7
soupsieve==2.0
sqlparse==0.3.1
SwampDragon==0.4.2.2
swapper==1.1.2.post1
text-unidecode==1.3
toml==0.10.0
tornado==6.0.4
tornado-redis==2.4.18
tqdm==4.47.0
twilio==6.38.1
twine==3.2.0
Twisted 
txaio==20.4.1
typed-ast==1.4.1
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
vine==1.3.0
virtualenv==20.0.31
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
whitenoise==5.0.1
wrapt==1.12.1
zope.interface==5.1.0
